refactor(storybook): split lib.rs into lib, scanner, helpers, and tests modules
The 1936-line crates/storybook/src/lib.rs was split into four modules. The
Scanner struct and its impl move to scanner.rs, free helper functions move
to helpers.rs, the tests move to tests.rs, and lib.rs keeps only the public
API surface (RULE_NAMES, Diagnostic types, options, scan_storybook). All
existing unit tests still pass. The Scanner impl is large enough that it
warrants a further per-topic split in a follow-up.
